Use stemplates for rendering

+++ templates/rollup-lib/codetemplate.py
@@ -1,9 +1,11 @@
 import os
 import pathlib
 import shutil
-from string import Template
 import json
+from stemplates import render_template
 
+
+# Template
 description = "Rollup Library"
 tags = ["javascript", "js", "library", "browser", "web", "rollup", "react", "npm"]
 
@@ -19,15 +21,15 @@ def ask_string(text: str) -> str:
 def open_stg_and_write_with(fin, fout, **kwargs):
     with open(fin, "r") as rc:
         tt = rc.read()
-        tp = Template(tt)
-        tt = tp.substitute(**kwargs)
+        tt = render_template(tt, **kwargs)
     with open(fout, "w") as ww:
         ww.write(tt)
 
 
 # TODO: Browser tests support using jest-puppeteer
 def new_project(where: pathlib.Path, resources_dir: pathlib.Path):
-    shutil.rmtree(where)
+    if os.path.exists(where):
+        shutil.rmtree(where)
     os.mkdir(where)
     os.chdir(where)
 
@@ -45,16 +47,23 @@ def new_project(where: pathlib.Path, resources_dir: pathlib.Path):
     os.system("npm init")
     ts_enabled = ask_boolean("Set up TypeScript?")
     tests_enabled = ask_boolean("Set up tests?")
-    deps = ["rollup", "@rollup/plugin-commonjs"]
+    deps = [
+        "rollup",
+        "@rollup/plugin-commonjs",
+        "cross-env",
+        "rollup-plugin-terser",
+        "rollup-plugin-node-resolve",
+        "del",
+    ]
     if ts_enabled:
         ts_deps = ["@rollup/plugin-typescript", "typescript"]
         deps = [*deps, *ts_deps]
     if tests_enabled:
         deps = [*deps, "jest"]
     if ts_enabled and tests_enabled:
-        deps = [*deps, "@types/jest"]
+        deps = [*deps, "@types/jest", "ts-jest"]
     deps_s = " ".join(deps)
-    # os.system(f"npm install {deps_s}")
+    os.system(f"npm install {deps_s}")
 
     # Create directories
     src_dir = ask_string("Directory name for library source code: ")
@@ -63,7 +72,7 @@ def new_project(where: pathlib.Path, resources_dir: pathlib.Path):
 
     # Modify package.json
     out_prefix = os.path.splitext(entry_file_name)[0]
-    package_json_path = os.path.join(src_dir, "package.json")
+    package_json_path = os.path.join("package.json")
     with open(package_json_path, "r") as pf:
         package_json = json.load(pf)
     package_json["main"] = f"./build/{out_prefix}.min.js"
@@ -73,14 +82,17 @@ def new_project(where: pathlib.Path, resources_dir: pathlib.Path):
         package_json["types"] = f"./build/{out_prefix}.d.js"
     package_json["files"] = ["build"]
     package_json["directories"] = {"test": "test"}
-    package_json["scripts"] = {
-        "build": "cross-env NODE_ENV=production rollup -c rollup.config.js",
-        "build-dev": "cross-env NODE_ENV=development rollup -c rollup.config.js",
-    }
+    package_json["scripts"] = {}
+    package_json["scripts"][
+        "build"
+    ] = "cross-env NODE_ENV=production rollup -c rollup.config.js"
+    package_json["scripts"]["build-dev"] = (
+        "cross-env NODE_ENV=development rollup -c rollup.config.js",
+    )
     if tests_enabled:
-        package_json["scripts"] = {"test": "jest"}
+        package_json["scripts"]["test"] = "jest"
     with open(package_json_path, "w") as pf:
-        json.dump(pf, package_json)
+        json.dump(package_json, pf, indent=4)
 
     # Create rollup config
     open_stg_and_write_with(
@@ -96,7 +108,7 @@ def new_project(where: pathlib.Path, resources_dir: pathlib.Path):
     # tsconfig.json
     if ts_enabled:
         open_stg_and_write_with(
-            fin=res_path("tsconfig.json.ctg"),
+            fin=res_path("tsconfig.json.stg"),
             fout="tsconfig.json",
             source_dir_path=src_dir,
         )
